Update addition checking algorithm
The way the algorithm would work is that it would push the new subsequence, then try obtaining the subseqeunce before the subsequence that was just append using some tricky methods.

With this commit, we first check if there are any additions, and once we're done, we then append a new subsequence.

This way, we also make this part of the code way more efficient and readable at once